Moved conceptual files to new package
