+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ConceptTests/Auto_IMU.java
@@ -6,7 +6,7 @@
 //
 // Also uses IMU to drive in a straight line when not avoiding an obstacle.
 
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.ConceptTests;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
