+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ConceptTests/Vuforia_PhoneCam.java
@@ -28,7 +28,7 @@
 
  */
 
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.ConceptTests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
